+++ src/pages/MainPage.tsx
@@ -1,6 +1,5 @@
 import { useState, useEffect } from "react";
 import { useLocation } from "react-router-dom";
-import { supabase } from "../../supabase/client";
 import { getDomainUrl } from "@/lib/utils";
 import {
   Carousel,
@@ -11,9 +10,8 @@ import {
 import useBaseStore from "@/stores/baseStore";
 import { Button } from "@/components/ui/button";
 import { analytics, analyticsTrack } from "@/analytics/analytics";
-import kakaoIcon from "@/assets/kakaoIcon.svg";
-import appleIcon from "@/assets/appleIcon.svg";
-import * as Sentry from "@sentry/react";
+import KakaoLoginBtn from "@/components/kakao/KakaoLoginBtn";
+import AppleLoginBtn from "@/components/auth/AppleLoginBtn";
 
 const MainPage: React.FC = () => {
   const user = useBaseStore((state) => state.user);
@@ -66,64 +64,18 @@ const MainPage: React.FC = () => {
     </div>
   );
 
-  const KakaoLoginBtn = () => {
+  const LoginBtnList = () => {
     const location = useLocation();
     const pathname = location.state?.from?.pathname || "";
-
     const pathParts = pathname.split("/");
-
     const groupId =
       pathParts.length === 3 && pathParts[1] === "group" ? pathParts[2] : "";
-
-    const handleKakaoLoginBtnClick = async () => {
-      analytics.track("클릭_카카오_로그인", { where: "KakaoLoginBtn" });
-      const { error } = await supabase.auth.signInWithOAuth({
-        provider: "kakao",
-        options: {
-          redirectTo: `${baseUrl}/term?groupId=${groupId}`,
-        },
-      });
-      if (error) {
-        console.error("Kakao login error:", error.message);
-        Sentry.captureException(error.message);
-      }
-    };
-
-    const handleAppleLoginBtnClick = async () => {
-      analytics.track("클릭_애플_로그인", { where: "AppleLoginBtn" });
-      const { error } = await supabase.auth.signInWithOAuth({
-        provider: "apple",
-        options: {
-          redirectTo: `${baseUrl}/term?groupId=${groupId}`,
-        },
-      });
-      if (error) {
-        console.error("Apple login error:", error.message);
-        Sentry.captureException(error.message);
-      }
-    };
+    const redirectUrl = `${baseUrl}/term?groupId=${groupId}`;
 
     return (
       <div className="flex flex-col gap-2">
-        <button
-          className="flex items-center gap-3 px-4 py-2 rounded-lg text-sm"
-          onClick={handleKakaoLoginBtnClick}
-          style={{ background: "#FEE500", color: "black" }}
-        >
-          <img src={kakaoIcon} className="w-4 h-4" />
-          카카오로 시작하기
-        </button>
-
-        {isIOSApp && (
-          <button
-            className="flex items-center gap-3 px-4 py-2 rounded-lg text-sm"
-            onClick={handleAppleLoginBtnClick}
-            style={{ background: "#222222", color: "white" }}
-          >
-            <img src={appleIcon} className="w-4 h-4" />
-            Apple로 시작하기
-          </button>
-        )}
+        <KakaoLoginBtn redirectUrl={redirectUrl} />
+        {isIOSApp &&  <AppleLoginBtn redirectUrl={redirectUrl} /> }
       </div>
     );
   };
@@ -218,7 +170,7 @@ const MainPage: React.FC = () => {
         <CarouselDots />
       </Carousel>
 
-      {user ? <PrayUStartBtn /> : <KakaoLoginBtn />}
+      {user ? <PrayUStartBtn /> : <LoginBtnList />}
     </div>
   );
 };
